@@ -32,16 +32,17 @@
     "node-uuid": "^1.4.8",
     "normalizr": "^3.2.2",
     "prettyjson": "^1.2.1",
-    "react": "^15.4.2",
+    "react": "^15.5.3",
     "react-bootstrap": "^0.30.8",
-    "react-dom": "^15.4.2",
+    "react-dom": "^15.5.3",
     "react-helmet": "^5.0.2",
     "react-redux": "^5.0.3",
-    "react-router": "^4.0.0",
+    "react-router": "^3.0.3",
     "react-router-bootstrap": "^0.23.1",
     "react-router-redux": "^4.0.8",
     "redux": "^3.6.0",
     "redux-devtools-extension": "^2.13.0",
+    "redux-form": "^6.6.2",
     "redux-thunk": "^2.2.0",
     "serve-favicon": "^2.4.2",
     "socket.io": "^1.7.3",
