@@ -54,8 +54,10 @@ app.use(bodyParser.json())
 
 app.use(express.static(__dirname + '../public'));
 
-// /api/todos/
-app.use(routes);
+//app.use(routes);
+app.use(routes.todos);
+app.use(routes.users);
+app.use(routes.counter);
 
 app.get('/*', (req, res) => {
     res.sendFile(path.join(__dirname, '..', 'index.html'));
