@@ -1,24 +1,24 @@
 import React from 'react'
 import { Provider } from 'react-redux'
 import { Router, Route, IndexRoute, browserHistory } from 'react-router'
-import MyDemo from './containers/'
+import MyApp from './containers/'
 //<Route path="/(:filter)" component={App}/>
 
 const router = (
-  <Route path="/" component={MyDemo.Home}>
-    <IndexRoute component={MyDemo.App}></IndexRoute>
-    <Route path="/about" component={MyDemo.About}></Route>
-    <Route path="/counter" component={MyDemo.Counter}></Route>
-    <Route path="/demo" component={MyDemo.Demo}></Route>
-    <Route path="/phone" component={MyDemo.Phone}></Route>
-    <Route path="/todos" component={MyDemo.Todos}></Route>
-    <Route path="/users" component={MyDemo.Users}></Route>
+  <Route path="/" component={MyApp.Home}>
+    <IndexRoute component={MyApp.App}></IndexRoute>
+    <Route path="/about" component={MyApp.About}></Route>
+    <Route path="/counter" component={MyApp.Counter}></Route>
+    <Route path="/contact" component={MyApp.Contact}></Route>
+    <Route path="/phone" component={MyApp.Phone}></Route>
+    <Route path="/todos" component={MyApp.Todos}></Route>
+    <Route path="/users" component={MyApp.Users}></Route>
   </Route>
 );
 
-const Root = ({ store, history }) => (
+const Root = ({ store }) => (
   <Provider store={store}>
-    <Router history={ history }>
+    <Router history={ browserHistory }>
       {router}
     </Router>
   </Provider>
