@@ -6,10 +6,6 @@ import reducers from './reducers'
 import { loadState, saveState } from './helpers/localStorage'
 import throttle from 'lodash/throttle';
 
-// fix location undefined issue.
-import createHistory from 'history/createBrowserHistory'
-import { ConnectedRouter, routerReducer, routerMiddleware, push } from 'react-router-redux'
-
 /*
  * if (window.devToolsExtension)
  * window.devToolsExtension.updateStore(store)
@@ -18,22 +14,12 @@ const configureStore = () => {
 
   const persistedState = loadState();
 
-
-  // Create an enhanced history that syncs navigation events with the store
-  const history = createHistory()
-
-  console.log('history', history);
-
-  //const middleware = routerMiddleware(history);
-
   //compose(applyMiddleware(thunk), devToolsEnhancer())
   const store = createStore(
     reducers,
     persistedState,
     compose(applyMiddleware(thunk), devToolsEnhancer())
   );
-
-  console.log('store', store);
 
   store.subscribe(throttle(() => {
     saveState({todos: store.getState().todos})
@@ -45,7 +31,7 @@ const configureStore = () => {
    * {"todos":[],"visibilityFilter":"SHOW_ALL"}
    */
 
-  return {store: store, history: history};
+  return store;
 };
 
 export default configureStore;
