@@ -12,7 +12,7 @@ class Home extends Component {
           <li><a href="/">App</a></li>
           <li><a href="/about">About</a></li>
           <li><a href="/counter">Counter</a></li>
-          <li><a href="/demo">Demo</a></li>
+          <li><a href="/contact">Contact</a></li>
           <li><a href="/phone">Phone</a></li>
           <li><a href="/todos">Todos</a></li>
           <li><a href="/users">Users</a></li>
